refactor: extract inline test modules to separate test files
Move the `#[cfg(test)] mod tests { ... }` blocks from seven source files into dedicated `*_tests.rs` files, using `#[path = "..."]` to keep the module structure unchanged. This reduces the main source files by over 1700 lines, making them easier to navigate while keeping tests discoverable in their own files.

Auto-committed-on: dragonfly
